+++ fhir_kindling/fhir_server/server_responses.py
@@ -1,5 +1,5 @@
 from typing import List
-from requests import Response
+from httpx import Response
 from fhir.resources.resource import Resource
 from fhir.resources.bundle import Bundle
 from fhir.resources.reference import Reference